requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
torch==2.1.0
torchvision==0.16.0
transformers==4.34.0
pillow==10.1.0
faiss-cpu==1.7.4
numpy==1.26.0
pandas==2.1.1
mlflow==2.7.1
prometheus-client==0.17.1
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
requests==2.31.0
pytest==7.4.2
httpx==0.25.0
black==23.9.1
flake8==6.1.0
tqdm==4.66.1
python-dotenv==1.0.0
